# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gruhnirman CRM API",
        "description": "API documentation for Gruhnirman CRM",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get all properties",
                "operationId": "e83a7fc5b1d98569892c10b7b0223879",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "builder_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "property_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Create a new property",
                "operationId": "70bf627c7a3b7593179b22a21956b41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "builder_id",
                                    "category_id",
                                    "property_type_id",
                                    "name",
                                    "starting_price"
                                ],
                                "properties": {
                                    "builder_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "property_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Sunrise Heights"
                                    },
                                    "starting_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4500000
                                    },
                                    "ending_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6000000
                                    },
                                    "image": {
                                        "description": "Property Image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, City"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "22.3421061"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "70.7299631"
                                    },
                                    "youtube_link": {
                                        "type": "string",
                                        "example": "https://youtu.be/..."
                                    },
                                    "brochure": {
                                        "description": "Property Brochure",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "additional_note": {
                                        "type": "string",
                                        "example": "Near Metro station"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "sq_feet",
                                                "price"
                                            ],
                                            "properties": {
                                                "sq_feet": {
                                                    "type": "string",
                                                    "example": "1000"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5000000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}